move service.h to include dir
